refactor(programareProcedurala/practica/types): updated the price per liter to be a float instead of a double

+++ programareProcedurala/practica/src/types/Beverages.hpp
@@ -15,7 +15,7 @@ struct Beverage {
     BeverageType type;
     std::string name;
     std::string color;
-    double pricePerLiter;
+    float pricePerLiter;
 };
 
 #endif
